move Zone to separate file
